Setup PostgreSQL Docker, configure EF Core, implement logging, restructuring
Docker was configured to start PostgreSQL using docker-compose. A docker-compose.yml file with the configuration for starting PostgreSQL server was created.

A DbContext was created to connect to the Entity Framework Core project that uses PostgreSQL as the database. The database connection string configuration is stored in user-secrets.

Logging of database queries executed by EF Core using ILoggerFactory has been enabled.

Entity Framework Core Snakify extension was used to format column names in the database in snake_case styles.

A custom implementation of the Result class was added.

A small restructuring of the project was done.